Almost there

+++ kernel/src/include/arch/arch_vmm.h
@@ -15,7 +15,7 @@ static void native_flush_tlb_single(unsigned long vaddr) {
 }
 
 static void native_flush_tlb_range(unsigned long vaddr,uint64 pages) {
-    for(int i = 0; i < pages; i++){
+    for(uint64 i = 0; i < pages; i++){
         asm volatile("invlpg (%0)" ::"r" (vaddr + (i * PAGE_SIZE)) : "memory");
     }
 }
